+++ app/src/main/java/com/wearerommies/roomie/presentation/navigator/MainNavigator.kt
@@ -8,8 +8,8 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navOptions
-import com.wearerommies.roomie.presentation.model.MainTabRoute
-import com.wearerommies.roomie.presentation.model.Route
+import com.wearerommies.roomie.presentation.navigator.route.MainTabRoute
+import com.wearerommies.roomie.presentation.navigator.route.Route
 import com.wearerommies.roomie.presentation.type.MainTabType
 import com.wearerommies.roomie.presentation.ui.home.navigateHome
 import com.wearerommies.roomie.presentation.ui.map.navigateMap
